Restructure notification
