+++ src/client/app/notifications/notification.service.ts
@@ -1,10 +1,12 @@
 import { Injectable, EventEmitter } from '@angular/core';
-import { AsyHttp, HttpOptions } from 'app/shared/asy-http.service';
-import { SocketService } from 'app/core/socket.service';
+
+import { AsyHttp, HttpOptions, PagingOptions } from 'app/shared';
+import { SocketService } from 'app/core';
+
 import { UserStateService } from 'app/admin/authentication/user-state.service';
-import { Notification } from './notification.class';
 import { UserService } from 'app/admin/users.service';
-import { PagingOptions } from 'app/shared/pager.component';
+
+import { Notification } from './notification.class';
 
 @Injectable()
 export class NotificationService {
