+++ src/client/app/notifications/notifications.module.ts
@@ -1,9 +1,10 @@
+import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
-import { NgModule } from '@angular/core';
+
+import { UtilModule } from 'app/shared';
 
 import { NotificationHandlerService } from './notification-handler.service';
-import { UtilModule } from 'app/shared/util.module';
 import { NotificationService } from './notification.service';
 import { NotificationIndicatorComponent } from './notification-indicator.component';
 import { NotificationStateService } from './notification-state-service';
